src/job_hunter_ai/drive/upload.py
"""
Google Drive upload functionality.

This module provides functions to upload generated documents to Google Drive.
Requires Google Drive API credentials.
"""

from pathlib import Path
from typing import Optional
import logging

# Import config
import sys
sys.path.insert(0, str(Path(__file__).parent.parent.parent))
from job_hunter_ai.config import (
    GOOGLE_DRIVE_FOLDER_ID,
    GOOGLE_CREDENTIALS_PATH,
)

logger = logging.getLogger(__name__)


class DriveUploadError(Exception):
    """Raised when Drive upload fails."""
    pass


def upload_to_drive(file_path: Path, folder_id: Optional[str] = None) -> str:
    """
    Upload a file to Google Drive.

    Args:
        file_path: Path to file to upload
        folder_id: Optional Drive folder ID (uses config default if None)

    Returns:
        Public URL to uploaded file

    Raises:
        DriveUploadError: If upload fails or credentials are missing

    TODO: Implement actual Google Drive API integration.
    This is a placeholder that will be implemented when Drive integration is needed.

    Future implementation will use:
    - google-api-python-client
    - service account credentials from GOOGLE_CREDENTIALS_PATH
    - folder ID from config or parameter
    """
    logger.warning(
        "Drive upload is not yet implemented. "
        "File would be uploaded: %s",
        file_path
    )

    # For now, return a placeholder URL
    # Real implementation would use google-api-python-client
    return f"https://drive.google.com/file/placeholder/{file_path.name}"


# Future implementation sketch:
"""
from google.oauth2 import service_account
from googleapiclient.discovery import build
from googleapiclient.http import MediaFileUpload

def upload_to_drive_impl(file_path: Path, folder_id: Optional[str] = None) -> str:
    '''Real implementation using Google Drive API.'''
    if not GOOGLE_CREDENTIALS_PATH:
        raise DriveUploadError("GOOGLE_CREDENTIALS_PATH not configured")

    credentials = service_account.Credentials.from_service_account_file(
        GOOGLE_CREDENTIALS_PATH,
        scopes=['https://www.googleapis.com/auth/drive.file']
    )

    service = build('drive', 'v3', credentials=credentials)

    file_metadata = {
        'name': file_path.name,
        'parents': [folder_id or GOOGLE_DRIVE_FOLDER_ID]
    }

    media = MediaFileUpload(str(file_path), resumable=True)

    file = service.files().create(
        body=file_metadata,
        media_body=media,
        fields='id, webViewLink'
    ).execute()

    return file.get('webViewLink')
"""
