requirements.txt
# Core LLM
groq>=0.4.0

# API and HTTP
requests>=2.31.0

# Google Sheets integration
gspread>=5.11.0
google-auth>=2.23.0

# Configuration and environment
PyYAML>=6.0.1
python-dotenv>=1.0.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: Google Drive upload (for future implementation)
# Uncomment when implementing Drive upload:
# google-api-python-client>=2.100.0
